tabla de usuarios, edit, new, etc.

+++ src/Controller/UserController.php
@@ -1,39 +1,35 @@
 <?php
 
 require_once './src/Repository/UserRepository.php';
-// require_once './Extensions/UserUtilitiesTrait.php';
 
-class UserController extends MainController{
+class UserController extends MainController {
 
-  public function viewLogin($error=array()) {
-    self::$twig->show('view_login.html', $error);
+  public function viewLogin($params = array("error" => '')) {
+    $this->render('login', $params);
   }
 
   public function login() {
-    if(!isset($_SESSION['id'])) { //no tiene ya una sesión iniciada
+    // if(!$this->isLoggedUser()) { //no tiene ya una sesión iniciada
       if(isset($_POST['username']) && isset($_POST['password'])){
         $username = $_POST['username'];
         $password = $_POST['password'];
         if(!empty($_POST['username']) && !empty($_POST['password'])){
           $repo = new UserRepository();
-          $username = $repo->login($username, $password);
-          if(!empty($username)) {
-            $this->startUserSession($username);
-            $this->redirectHome();
+          $user = $repo->login($username, $password);
+          if(!empty($user)) {
+            $this->startUserSession($user);
+            $_GET['action'] = ''; 
+            $this->dispatch('users');
           }else{
-            echo('Usuario o contraseña incorrectos');
-            //devolver a Login con el error
-            // $this->viewLogin(array("error"=>'Usuario o contraseña incorrectos'));
+            $this->render('login', array("error" => 'Usuario o contraseña incorrectos'));
           }
         }else {
-          echo('Faltó completar alguno de los datos');
-          // $this->viewLogin(array("error"=>'Faltó completar alguno de los datos'));
+          $this->render('login', array("error" => 'Faltó completar alguno de los datos'));
         }
       }else {
-        echo('Faltó completar alguno de los datos');
-        // $this->viewLogin(array("error"=>'Faltó completar alguno de los datos'));
+        $this->render('login', array("error" => 'Faltó completar alguno de los datos'));
       }
-    }
+    // }
   }
 
   private function startUserSession($user){
@@ -41,18 +37,22 @@ class UserController extends MainController{
     session_set_cookie_params(0);
     session_start();
     $_SESSION['id']= $user['id'];
-    $_SESSION['user']= $user['usuario'];
+    $_SESSION['username']= $user['username'];
   }
 
   public function logout(){
     session_destroy();
     $_SESSION= array();
-    $this->redirectHome();
+    $this->dipatch('home');
   }
 
-  public function register() {
+  public function viewUserNew($params = array("error" => '')) {
+    $this->render('user_new', $params);
+  }
 
-    if(isset($_SESSION['id'])){//es un usuario logueado
+  public function userNew() {
+
+    if($this->isLoggedUser()) { //es un usuario logueado
 
       $lastName = isset($_POST['last_name']) ? $_POST['last_name'] : '';
       $firstName =  isset($_POST['first_name']) ? $_POST['first_name'] : '';
@@ -79,66 +79,58 @@ class UserController extends MainController{
                 $password, 
                 $username
               );
-              echo('El usuario fue agregado exitosamente');
-              // $this->viewUsersList('success', 'El usuario fue agregado exitosamente');
+              $this->dispatch('users');
             }else {
-              echo('Se produjo un error: el email ingresado ya existe');
-              // $this->viewUsersList('error', 'Se produjo un error: el email ingresado ya existe');
+              $this->render('user_new', array('error' => 'Se produjo un error: el email ingresado ya existe'));
             }
           }else {
-            echo('Se produjo un error: el nombre de usuario ingresado ya existe');
-            // $this->viewUsersList('error', 'Se produjo un error: el nombre de usuario ingresado ya existe');
+            $this->render('user_new', array('error' => 'Se produjo un error: el nombre de usuario ingresado ya existe'));
           }
       }else {
-        echo($err);
-        // $this->viewUsersList('error', $err);
+        $this->render('user_new', array('error' => $err));
       }    
     }
 
   }
 
-  public function user(){
-    // if(isset($_SESSION['id'])){//es un usuario logueado
-      $_GET['action']='';
+  public function user($data = array("error" => '')) {
+    if($this->isLoggedUser()) { //es un usuario logueado
       $user_repo = new UserRepository();
       $user_id = $_POST["id"];
-      $data = array();
       $data['user'] = $user_repo->getUser($user_id);
-      if(!empty($data)){
-          echo(json_encode($data));
+      if (!empty($data)){
+        $this->render('user_edit', $data);
       }else {
-        echo('error');
-        // $this->viewUsersList('error', 'Se produjo un error');
+        $data['error'] = 'No existe el usuario';
+        $this->render('user_edit', $data);
       }
-    // }else {
-    //   echo('Debe iniciar sesión');
-    // }
+    }else {
+      $this->render('view_login');
+    }
   }
 
   public function users($state=NULL, $msg="") {
-    // if(isset($_SESSION['id'])){//es un usuario logueado
+    if($this->isLoggedUser()) { //es un usuario logueado
 
-        $user_repo = new UserRepository();
-        $data = array();
+      $user_repo = new UserRepository();
+      $data = array();
 
-        if(!is_null($state)){
-          $data[$state]= $msg;
-        }
+      if(!is_null($state)){
+        $data[$state]= $msg;
+      }
 
-        $users= $user_repo->getAllUsuarios();
-        $data['users']= $users;
+      $users= $user_repo->getUsers();
+      $data['users'] = $users;
 
-        echo(json_encode($data));
-        die;
-        // $this::$twig->show('list_users.html', $data);
-    // }else{
-      $this->redirectHome();
-    // }
+      $this->render('users', $data);
+    }else{
+      $this->render('view_login');
+    }
   }
 
   public function userUpdate(){
-    // if(isset($_SESSION['id'])){ //es un usuario logueado
-      $userId = isset($_POST['user_id']) ? $_POST['user_id'] : '';
+    if ($this->isLoggedUser()) { 
+      $userId = isset($_POST['id']) ? $_POST['id'] : '';
       $lastName = isset($_POST['last_name']) ? $_POST['last_name'] : '';
       $firstName =  isset($_POST['first_name']) ? $_POST['first_name'] : '';
       $email = isset($_POST['email']) ? $_POST['email'] : '';
@@ -152,13 +144,17 @@ class UserController extends MainController{
         $password, 
         $username
       );
-    
-      if(empty($err)){
-        $user_repo= new UserRepository();
-        if ($user_repo->getUser($userId)) {
 
-          if($user_repo->checkUserName($username)){
-            if($user_repo->checkEmail($email)) {
+      $user_repo= new UserRepository();
+
+      $user = $user_repo->getUser($userId);
+
+      if (empty($err)) {
+
+        if ($user) {
+
+          if($user_repo->checkUserName($username, $userId)){
+            if($user_repo->checkEmail($email, $userId)) {
 
               $user = $user_repo->updateUser(
                 $userId,
@@ -169,44 +165,59 @@ class UserController extends MainController{
                 $username
               );
 
-              echo('Usuario actualizado.');
-              // $this->viewUsersList('success', 'El usuario se actualizó exitosamente');
-                
+              $this->dispatch('users');
+              
+            }else {
+              $this->render('user_edit',
+                array(
+                  'error' => 'Se produjo un error: el email ingresado ya existe',
+                  'user' => $user
+                )
+              );
+            }
           }else {
-            echo('Se produjo un error: el email ingresado ya existe');
-            // $this->viewUsersList('error', 'Se produjo un error: el email ingresado ya existe');
+            $this->render('user_edit', 
+              array(
+                'error' => 'Se produjo un error: el nombre de usuario ingresado ya existe',
+                'user' => $user
+              )
+            );
           }
         }else {
-          echo('Se produjo un error: el nombre de usuario ingresado ya existe');
-          // $this->viewUsersList('error', 'Se produjo un error: el nombre de usuario ingresado ya existe');
-        }
+          $this->render('user_edit', 
+            array(
+              'error' => 'Se produjo un error: No existe el usuario',
+              'user' => $user
+            )
+          );
+        }   
       }else {
-        echo('Se produjo un error: no existe el usuario');
-      }   
-    }else {//error con algun campo del form
-      echo($err);
-      // $this->viewUsersList('error', $err);
+        $this->render('user_edit', 
+          array(
+            'error' => $err,
+            'user' => $user
+          )
+        );
+      } 
+    }else {
+      $this->render('view_login');
     }
-      
-    // }else {
-    //   $this->redirectHome();
-    // }
   }
 
   public function userDelete() {
-    // if(isset($_SESSION['id'])){ //es un usuario logueado
-      // if($_POST['id_user'] != $_SESSION['id']){ //el que quiere eliminar no es él mismo
+    if($this->isLoggedUser()){ //es un usuario logueado
+      if($_POST['id_user'] != $_SESSION['id']){ //el que quiere eliminar no es él mismo
         $user_repo= new UserRepository();
         $user_repo->removeUser($_POST['id_user']);
         echo('El usuario fue eliminado');
         // $this->viewUsersList('success', 'El usuario fue eliminado');
-      // }else{
-      //   echo('Se produjo un error: no puedes eliminar a ese usuario');
-      //   // $this->viewUsersList('error', 'Se produjo un error: no puedes eliminar a ese usuario');
-      // }
-    // }else {
-    //   $this->redirectHome();
-    // }
+      }else{
+        echo('Se produjo un error: no puedes eliminar a ese usuario');
+        // $this->viewUsersList('error', 'Se produjo un error: no puedes eliminar a ese usuario');
+      }
+    }else {
+      $this->render('view_login');
+    }
 
   }
 
@@ -241,10 +252,10 @@ class UserController extends MainController{
       if (!filter_var($email, FILTER_VALIDATE_EMAIL) === true) {
         $err.= 'error en email: no es un email válido. ';
       }
-      if ( ( strlen($username) < 5 ) || ( strlen($username) > 20 ) || (!ctype_alnum($username) ) ){
+      if ( ( strlen($username) < 6 ) || ( strlen($username) > 20 ) || (!ctype_alnum($username) ) ){
         $err.= 'error en nombre de usuario: mínimo 6 caracteres alfanuméricos, máximo 20. ';
       }
-      if ( ( strlen($password) < 8 ) || ( strlen($password) > 20 ) ) {
+      if ( ( strlen($password) < 6 ) || ( strlen($password) > 20 ) ) {
         $err.= 'error en password: mínimo 8 caracteres, máximo 20. ';
       }
     }else {
